requirements.txt
streamlit>=1.31.0
celery>=5.3.4
redis>=5.0.1
google-generativeai>=0.3.2
Pillow>=10.0.0
pandas==2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
instagrapi==2.0.0
schedule==1.2.0
python-multipart==0.0.6
fastapi==0.104.0
uvicorn==0.24.0
cloudinary>=1.36.0
sqlalchemy>=2.0.0
plotly>=5.0.0
python-dateutil>=2.8.2
matplotlib>=3.8.0
seaborn>=0.13.0
pytz>=2024.1
